User delete corrective

+++ src/Controller/UserController.php
@@ -85,8 +85,9 @@ class UserController extends AbstractController
     /**
      * @Route("/users/{id}/delete", name="user_delete")
      */
-    public function taskDelete(User $user)
+    public function userDelete(User $user)
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
         $em = $this->getDoctrine()->getManager();
         $em->remove($user);
         $em->flush();
